refactor(inputs): extract input diagnostic helpers
